Revamped structure and improved algos
