+++ pyfilter/__init__.py
@@ -1,1 +1,1 @@
-__version__ = '0.3.20'
+__version__ = '0.4.1'
