+++ pyfilter/algorithms/kernels/__init__.py
@@ -1,0 +1,3 @@
+from .online import ShrinkingKernel, NonShrinkingKernel, KernelDensitySampler, AdaptiveKernel, OnlineKernel
+from .mh import SymmetricMH, ApproximateSymmetricMH, ParticleMetropolisHastings
+from .base import BaseKernel
