+++ pyfilter/algorithms/ness.py
@@ -1,63 +1,26 @@
-from .base import SequentialAlgorithm
-from ..filters.base import ParticleFilter, cudawarning
-from ..utils import get_ess, normalize
-import torch
-from ..resampling import residual
-from .kernels import AdaptiveKernel
+from .base import SequentialParticleAlgorithm
+from ..utils import get_ess
+from .kernels import AdaptiveKernel, OnlineKernel
 from ..filters import SISR
 
 
-class NESS(SequentialAlgorithm):
-    def __init__(self, filter_, particles, threshold=0.9, resampling=residual, kernel=None):
+class NESS(SequentialParticleAlgorithm):
+    def __init__(self, filter_, particles, kernel=None):
         """
-        Implements the NESS alorithm by Miguez and Crisan.
-        :param particles: The particles to use for approximating the density
-        :type particles: int
-        :param threshold: The threshold for when to resample the parameters
-        :type threshold: float
+        Implements the NESS algorithm by Miguez and Crisan.
         :param kernel: The kernel to use when propagating the parameter particles
-        :type kernel: pyfilter.algorithms.kernels.BaseKernel
+        :type kernel: OnlineKernel
         """
-
-        cudawarning(resampling)
 
         if isinstance(filter_, SISR) and filter_._th != 1.:
             raise ValueError('The filter must have `ess = 1.`!')
 
-        super().__init__(filter_)
+        super().__init__(filter_, particles)
 
-        self._kernel = kernel or AdaptiveKernel(ess=threshold, resampling=resampling)
-        self._filter.set_nparallel(particles)
+        self._kernel = kernel or AdaptiveKernel()
 
-        # ===== Weights ===== #
-        self._w_rec = torch.zeros(particles)
-
-        # ===== ESS related ===== #
-        self._logged_ess = (torch.tensor(particles, dtype=self._w_rec.dtype),)
-        self._particles = particles
-
-    def initialize(self):
-        """
-        Overwrites the initialization.
-        :return: Self
-        :rtype: NESS
-        """
-
-        self.filter.ssm.sample_params(self._particles)
-
-        shape = (self._particles, 1) if isinstance(self.filter, ParticleFilter) else (self._particles,)
-        self.filter.viewify_params(shape).initialize()
-
-        return self
-
-    @property
-    def logged_ess(self):
-        """
-        Returns the logged ESS.
-        :rtype: torch.Tensor
-        """
-
-        return torch.tensor(self._logged_ess)
+        if not isinstance(self._kernel, OnlineKernel):
+            raise ValueError('Kernel must be of instance {}!'.format(OnlineKernel.__class__.__name__))
 
     def _update(self, y):
         # ===== Jitter ===== #
@@ -72,17 +35,3 @@ class NESS(SequentialAlgorithm):
         self._logged_ess += (ess,)
 
         return self
-
-    def predict(self, steps, aggregate=True, **kwargs):
-        px, py = self.filter.predict(steps, aggregate=aggregate, **kwargs)
-
-        if not aggregate:
-            return px, py
-
-        w = normalize(self._w_rec)
-        wsqd = w.unsqueeze(-1)
-
-        xm = (px * (wsqd if self.filter.ssm.hidden_ndim > 1 else w)).sum(1)
-        ym = (py * (wsqd if self.filter.ssm.obs_ndim > 1 else w)).sum(1)
-
-        return xm, ym
